+++ VTK/Utilities/tiff/tif_dirwrite.c
@@ -1,4 +1,4 @@
-/* $Header: /cvsroot/ParaView3/ParaView3/VTK/Utilities/tiff/Attic/tif_dirwrite.c,v 1.1 2002-01-29 22:38:20 andy Exp $ */
+/* $Header: /cvsroot/ParaView3/ParaView3/VTK/Utilities/tiff/Attic/tif_dirwrite.c,v 1.2 2002-06-24 22:02:59 andy Exp $ */
 
 /*
  * Copyright (c) 1988-1997 Sam Leffler
@@ -564,7 +564,7 @@ TIFFWritePerSampleShorts(TIFF* tif, ttag_t tag, TIFFDirEntry* dir)
         uint16* w = buf;
         int i, status, samples = tif->tif_dir.td_samplesperpixel;
 
-        if (samples > NITEMS(buf))
+        if (samples > (int)NITEMS(buf))
                 w = (uint16*) _TIFFmalloc(samples * sizeof (uint16));
         TIFFGetField(tif, tag, &v);
         for (i = 0; i < samples; i++)
@@ -589,7 +589,7 @@ TIFFWritePerSampleAnys(TIFF* tif,
         int i, status;
         int samples = (int) tif->tif_dir.td_samplesperpixel;
 
-        if (samples > NITEMS(buf))
+        if (samples > (int)NITEMS(buf))
                 w = (double*) _TIFFmalloc(samples * sizeof (double));
         TIFFGetField(tif, tag, &v);
         for (i = 0; i < samples; i++)
